requirements.txt
google-genai
jinja2
streamlit
python-dotenv